requirements.txt
openstacksdk
click
logging